experiment-generator/requirements.txt
Click==7.0
pymongo==3.10.1
python-dotenv==0.11.0
redis==3.4.1
rq==1.2.2
